requirements.txt
Django==1.5.4
Markdown==2.3.1
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
django-simple-friends==1.0.0
django-toolbelt==0.0.1
djangorestframework==2.3.8
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
